+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Change: Add `$adjacent` argument to a number of filters, to sync with WP core
     * Adds to `{$adjacent}_post_link`, `{$adjacent}_or_loop_post_link`, `c2c_{$adjacent}_or_loop_post_link_get`, and `c2c_{$adjacent}_or_loop_post_link_output`
 * Change: Use `c2c_get_adjacent_or_loop_post()` to obtain post, rather than duplicating its functionality in `c2c_get_adjacent_or_loop_post_link()`
+* Change: Remove unnecessary (and incorrect) determination of adjacent post in `c2c_adjacent_or_loop_post_link()`
 * Change: Use a different variable name to avoid changing variable sent as function argument and later passed as argument to filter
 * Change: Initialize plugin on 'plugins_loaded' action instead of on load
 * Change: Update code formatting to match modern WordPress standards